Structured logging events 2: remove global logger lock (#3795)
The global logger created by `init_logger` is currently behind a `RwLock`, which must be acquired on every logging operation. The only reason for the existence of this lock, is to support the testing scenario where the stdout/stderr loggers needs to be substituted to write to different file targets.

This PR removes the global logger lock by shifting the burden of stdout/stderr entirely into the test framework, who does so by redirecting the file handles at the OS level.

* logger: remove global lock
* `IoArgs`: remove `stdout` and `stderr`
* goldie: use unified diff for test failure output

GitOrigin-RevId: 0726822b2c3c541fc05a03bcd264eb7d67ef03ba