Remove frontend schemas from dbt-fusion
The `dbt-frontend-schemas` was placed in the SA `dbt-fusion`, but it's
generally unused there. The only used components is `Ident` /
`Identifier`, which was only opportunistically part of
`dbt-frontend-schemas` -- it belongs more to `dbt-frontend-common`, but
was placed in "schemas" so that `make-functions` doesn't depend on arrow
or datafusion and is quick to compile on CI.

This commit

- moves `dbt-frontend-schemas` out of SA `dbt-fusion`, as it is unused
  there
- the `Ident` / `Identifier` is moved to a small new crate (`dbt-ident`), so
  that "common" and "schemas" can share it without depending on each other
  (which is now problematic) and so that `make-functions` is still fast
  to compile
- few dialect-related usages are updated. They fall into two categories
  1. they used "schemas" Dialect where "common" Dialect would be more
     appropriate
  2. they map between "schemas" and "common" Dialects.

In the future, the two Dialect enums can be combined into one and shared
(just like `Ident` / `Identifier`), avoiding the need for any type
back-and-forth on the dialect.

GitOrigin-RevId: b877ad55ce42202ad3d8506ba49658b76e057d19