Fix failure when computing Snowflake distinct array_agg
This changes how Snowflake array_agg is implemented.
Before the change it was implemented via 'simplify', but this has the
limitation that aggregated values need to be the same as sorting values
when distinct is on (otherwise results would not be well-defined). The
new implementation delegates at the accumulator level and casts to
VARIANT at the aggregation finalize stage.

For this to work, some changes are backported in DataFusion fork.

GitOrigin-RevId: 5bf1370d844aba5df3fa59d7ddc2b5a4efc84bab