Improve compat with `dbt-external-tables` (#5129)
* Preserve `DbtSource.other` in `resolve_sources()`

Some keys like `sources.tables.external` are "externally typed", i.e dbt
does not define an explicit schema for them, and users can implement
their own packages that consume the values in that key.

One such package is `dbt-external-tables`, which accepts `location` and
some other parameters.

This commit makes sure we're preserving the value defined in their
project configs when generating the manifest. These values are held in
the `other` key of `DbtSource`, and they're passed down to any macros
that use it.

* Make `dbt run-operation stage_external_sources` work

The macro `stage_external_sources` is widely used in
`dbt-external-tables`, and it relies on `flags.PRINTER_WIDTH` and
`ResultObject.keys()`.

This commit implements those so that the macro does not break.

* Changie

GitOrigin-RevId: 62be93f152f48f25ffd543cb942c10e51ef0bbfe